+++ example/android/app/src/main/java/com/androidwidgetexample/WidgetConfigurationActivity.java
@@ -1,0 +1,6 @@
+package com.androidwidgetexample;
+
+import com.reactnativeandroidwidget.RNWidgetConfigurationActivity;
+
+public class WidgetConfigurationActivity extends RNWidgetConfigurationActivity {
+}