+++ example/index.ts
@@ -1,7 +1,11 @@
 import { AppRegistry } from 'react-native';
-import { registerWidgetTaskHandler } from 'react-native-android-widget';
+import {
+  registerWidgetConfigurationScreen,
+  registerWidgetTaskHandler,
+} from 'react-native-android-widget';
 import { name as appName } from './app.json';
-import { App, widgetTaskHandler } from './src';
+import { App, WidgetConfigurationScreen, widgetTaskHandler } from './src';
 
 AppRegistry.registerComponent(appName, () => App);
 registerWidgetTaskHandler(widgetTaskHandler);
+registerWidgetConfigurationScreen(WidgetConfigurationScreen);