+++ example/src/index.ts
@@ -1,2 +1,3 @@
 export * from './App';
+export * from './WidgetConfigurationScreen';
 export * from './widgetTaskHandler';