+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to configure (and reconfigure) widget once it is added on home screen
+
 ## [0.7.0] - 2023-07-15
 
 ### Added